+++ Lection1_HW/hw1.py
