+++ Lection2_HW/hw1.py
