+++ Lection2_HW/hw2.py
